@@ -1,10 +1,12 @@
 {
   "name": "root",
+  "type": "module",
   "private": true,
   "packageManager": "pnpm@7.30.0",
   "scripts": {
     "start": "pnpm -C packages/playground dev",
-    "build": "build"
+    "build": "build",
+    "lint": "eslint ."
   },
   "devDependencies": {
     "@antfu/eslint-config": "^2.6.1",
@@ -20,14 +22,16 @@
     "@types/js-beautify": "^1.13.3",
     "@types/lodash-es": "^4.17.5",
     "@types/node": "^16.11.6",
-    "@vitejs/plugin-vue": "^2.3.3",
-    "@vitejs/plugin-vue-jsx": "^1.3.10",
-    "autoprefixer": "^10.4.16",
+    "@unocss/eslint-config": "^0.58.3",
+    "@vitejs/plugin-vue": "5.0.2",
+    "@vitejs/plugin-vue-jsx": "3.1.0",
     "cross-env": "^7.0.3",
-    "eslint": "^7.32.0",
+    "eslint": "^8.36.0",
     "eslint-config-prettier": "^8.10.0",
+    "eslint-plugin-format": "^0.1.0",
     "eslint-plugin-prettier": "^4.2.1",
     "lint-staged": "^11.2.6",
+    "lodash-es": "^4.17.21",
     "prettier": "^2.4.1",
     "rimraf": "^3.0.2",
     "simple-git-hooks": "^2.9.0",
@@ -36,11 +40,16 @@
     "stylelint-config-standard": "^22.0.0",
     "stylelint-order": "^5.0.0",
     "stylelint-scss": "^3.21.0",
-    "tsup": "^6.1.3",
-    "typescript": "^4.4.3",
-    "vite": "^2.9.13"
+    "tsup": "^6.7.0",
+    "typescript": "^5.0.2",
+    "unocss": "^0.58.3",
+    "vite": "5.0.10",
+    "vue-tsc": "^1.8.27"
   },
   "simple-git-hooks": {
     "pre-commit": "npx lint-staged"
+  },
+  "lint-staged": {
+    "*": "eslint --fix"
   }
 }
